LayoutTests/platform/ios-simulator/editing/pasteboard/block-wrappers-necessary-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldDeleteDOMRange:range from 1 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document to 86 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of P > CENTER > DIV > DIV > BODY > HTML > #document to 2 of CENTER > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of P > CENTER > DIV > DIV > BODY > HTML > #document to 0 of P > CENTER > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldInsertNode:#document-fragment replacingDOMRange:range from 0 of P > CENTER > DIV > DIV > BODY > HTML > #document to 0 of P > CENTER > DIV > DIV > BODY > HTML > #document givenAction:WebViewInsertActionPasted
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 2 of CENTER > DIV > DIV > BODY > HTML > #document to 2 of CENTER > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document to 85 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldApplyStyle:<DOMCSSStyleDeclaration: 0x7fce13e8b590> toElementsInDOMRange:range from 0 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document to 85 of #text > P > CENTER > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 1 of P > CENTER > DIV > DIV > BODY > HTML > #document to 85 of #text > SPAN > FONT > P > CENTER > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of #text > SPAN > FONT > DIV > P > CENTER > DIV > DIV > BODY > HTML > #document to 85 of #text > SPAN > FONT > P > CENTER > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification


Tests: 

Bug fix for <rdar://problem/4180820> Erroneous spacing between lines introduced on paste

Applying block styles, like text-align, used to wrap paragraphs that needed block style changes in clones of other paragraphs in the document.  This would almost always introduce unwanted styles.


Before cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       "
<#selection-anchor>This is an "
|       <b>
|         "interactive"
|       " documentation site, so please help out by posting any useful information that you've had to search out."
|       <br>
|       "
(Adding to this "
|       <a>
|         href="db.cgi?Wiki_Is_So_Easy"
|         "wiki is so easy"
|       ", you can't screw it up). See "
|       <a>
|         href="db.cgi?Wiki_Reference"
|         "Wiki Reference"
|       " for text formatting syntax."
|       <br>
|       "
Info for setting up your own wiki is at "
|       <a>
|         href="db.cgi?About_Howto.Apple.Com"
|         "about howto.apple.com"
|       "."
|       <br>
|       "
Check out a cool wiki editor application that helps in wiki generation: "
|       <a>
|         href="db.cgi?Wikinator"
|         "Wikinator"
|       "
"
|     <p>
|       "
Start browsing the topics below, or use the search field at the bottom of every page.<#selection-focus>
"
|   "
"
| "
"

After cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       <#selection-caret>
|       <br>
|   "
"
| "
"

After paste (this should be identical to before cut):
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <center>
|     "
"
|     <p>
|       <div>
|         style="text-align: start;"
|         <font>
|           size="3"
|           <span>
|             style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|             "This is an "
|             <b>
|               "interactive"
|             " documentation site, so please help out by posting any useful information that you've had to search out."
|       <div>
|         style="text-align: start;"
|         <font>
|           size="3"
|           <span>
|             style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|             "(Adding to this "
|             <a>
|               href="db.cgi?Wiki_Is_So_Easy"
|               "wiki is so easy"
|             ", you can't screw it up). See"
|             " "
|             <a>
|               href="db.cgi?Wiki_Reference"
|               "Wiki Reference"
|             " for text formatting syntax."
|       <div>
|         style="text-align: start;"
|         <font>
|           size="3"
|           <span>
|             style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|             "Info for setting up your own wiki is at "
|             <a>
|               href="db.cgi?About_Howto.Apple.Com"
|               "about howto.apple.com"
|             "."
|       <div>
|         style="text-align: start;"
|         <font>
|           size="3"
|           <span>
|             style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|             "Check out a cool wiki editor application that helps in wiki generation: "
|             <a>
|               href="db.cgi?Wikinator"
|               "Wikinator"
|     <p>
|       style="text-align: start;"
|       <font>
|         size="3"
|         <span>
|           style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|           "Start browsing the topics below, or use the search field at the bottom of every page.<#selection-caret>"
|   "
"
| "
"
